requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

# --- Web API ---
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.3.4
python-multipart==0.0.9

# --- Numerika ---
numpy==1.26.4

# --- Embeddings + QA ---
sentence-transformers==3.0.1
transformers==4.44.2
torch==2.3.1+cpu

# FAISS + BM25 + tokenizers
faiss-cpu==1.12.0
rank-bm25==0.2.2
sentencepiece==0.1.99

# --- PDF + OCR ---
PyMuPDF==1.24.9
easyocr==1.7.1
Pillow==10.4.0
opencv-python-headless==4.10.0.84
scikit-image==0.24.0

# --- NER ---
spacy==3.6.1   # (uklonjen duplikat 3.7.5)

# --- Testovi / dev ---
pytest==8.3.2
httpx==0.27.2
reportlab==4.2.2
protobuf>=4.25,<5
